(wip) Implementing collaborational features. This specific commit relates to a `find` view and tabular recipients layout.
>>
The idea is to allow the gift coordinator to 'invite' a recipient to an event, where that user can accept/decline the invitation. The gifter may also not want to notify the user of adding them to the event, so that user will be added as a recipient to the event without ever being notified/added to the event on their end.